+++ libs/dialyzer/erts_internal.erl
@@ -28,7 +28,12 @@
 
 -export([
     cmp_term/2,
-    atomics_new/2
+    atomics_new/2,
+    counters_new/1,
+    counters_get/2,
+    counters_add/3,
+    counters_put/3,
+    counters_info/1
 ]).
 
 -spec cmp_term(A :: term(), B :: term()) -> -1 | 0 | 1.
@@ -39,4 +44,28 @@ cmp_term(_A, _B) ->
 %% atomics:new/2 in libs/estdlib), not the user-facing options list.
 -spec atomics_new(Arity :: pos_integer(), OptsBitmask :: non_neg_integer()) -> reference().
 atomics_new(_Arity, _OptsBitmask) ->
+    erlang:nif_error(undefined).
+
+-spec counters_new(Size :: pos_integer()) -> reference().
+counters_new(_Size) ->
+    erlang:nif_error(undefined).
+
+-spec counters_get(Ref :: reference(), Ix :: pos_integer()) -> integer().
+counters_get(_Ref, _Ix) ->
+    erlang:nif_error(undefined).
+
+-spec counters_add(Ref :: reference(), Ix :: pos_integer(), Incr :: integer()) -> ok.
+counters_add(_Ref, _Ix, _Incr) ->
+    erlang:nif_error(undefined).
+
+-spec counters_put(Ref :: reference(), Ix :: pos_integer(), Value :: integer()) -> ok.
+counters_put(_Ref, _Ix, _Value) ->
+    erlang:nif_error(undefined).
+
+-spec counters_info(Ref :: reference()) ->
+    #{
+        size := non_neg_integer(),
+        memory := non_neg_integer()
+    }.
+counters_info(_Ref) ->
     erlang:nif_error(undefined).